feat(watchlist): add POST /watchlist/bulk endpoint
Bulk-creates watchlist entries with server-side dedup by (user_id, symbol).
Mirrors the alerts/holdings bulk pattern: validation failures abort with 422
before any insert; duplicates are pre-filtered against existing rows and
within-batch repeats; insert uses sqlite_insert with on_conflict_do_nothing
to match the single-add code path.

Also fixes the watchlist id column from BigInteger to Integer to match every
other table — BigInteger is not aliased to ROWID on SQLite so ORM-side
inserts hit NOT NULL on id. The production DB schema is already INTEGER, so
this is a model-side alignment with zero on-disk impact.

+++ fin/models/watchlist.py
@@ -1,6 +1,6 @@
 from datetime import datetime, timezone
 
-from sqlalchemy import BigInteger, Column, DateTime, String, UniqueConstraint
+from sqlalchemy import BigInteger, Column, DateTime, Integer, String, UniqueConstraint
 
 from fin.database import Base
 
@@ -13,7 +13,7 @@ class WatchlistModel(Base):
         UniqueConstraint("user_id", "symbol", name="uq_watchlist_user_symbol"),
     )
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     user_id = Column(BigInteger, nullable=True)
     symbol = Column(String, nullable=False, index=True)
     name = Column(String)
